Move stat struct into stat.go
